+++ js/data.js
@@ -1,0 +1,37 @@
+const TIMES = ['12:00', '13:00', '14:00'];
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+const TITLE = 'Трешка в центре';
+const OFFER_COUNT = 10;
+const PRICE_MIN = 100;
+const PRICE_MAX = 100;
+const GUESTS_MAX = 10;
+const ROOMS_MAX = 10;
+const DESCRIPTION =
+  'Сдаём нашу любимую квартиру в связи с необходимостью расширения. Удобная кухня - гостиная 14,5 м² остаётся новым хозяевам со всей мебелью и техникой. Также оставим мебель в коридоре и спальне, а также технику в ванной. Квартира удачно расположена на втором этаже , на первом прямо под квартирой колясочная зона.';
+
+export {
+  TIMES,
+  TYPES,
+  FEATURES,
+  PHOTOS,
+  TITLE,
+  OFFER_COUNT,
+  PRICE_MAX,
+  PRICE_MIN,
+  GUESTS_MAX,
+  ROOMS_MAX,
+  DESCRIPTION,
+};
